PG-1968 Migrate _keys file to the new format
This adds `pg_tde/keys_version`, which contains the latest magic number
(version) of the existing wal and smgr keys. If those versions don't
match the current ones at the server start, it will rewrite _keys to
the new format and update keys_version.
If a crash happens wile updating _keys, it will retry again on the next
start since keys_version wasn't updated.

We extract the version number from FILE_MAGIC. And define a function
for each previous version of the keys file to transform old records
into the current format. This way, introducing changes into the keys
format, we just bump the numeric version of FILE_MAGIC. And
define/update transforming functions for all the previous formats
(and add them to `pg_tde_migrate_smgr/wal_keys_file` functions).

With the current FILE_MAGIC size we are restricted to 255 file
versions, but let's solve that when (if) we get there.